32-bit data table for C

+++ header.h
@@ -4,7 +4,7 @@
 extern "C" {
 #endif
 
-extern void (*const core_1brc)(int32_t,uint64_t*,uint32_t,int8_t*,uint32_t*,int64_t*,int8_t*,uint64_t);
+extern void (*const core_1brc)(int32_t,uint64_t*,uint32_t,int8_t*,uint32_t*,int32_t*,int8_t*,uint64_t);
 extern uint64_t (*const core_1brc_buf_elts)();
 extern uint64_t (*const core_1brc_periter)();
